Implement a minimal viable version of expand
